replaced DB usage in lambas with REST api + timeout bugfix + improve documents & scripts for rest-api-client generation

+++ packages/rest-api-client/generate.sh
@@ -1,5 +1,8 @@
+
+if [ -z $API_URI ]; then echo "API_URI is missing"; exit; fi
+
 docker run --rm -v "${PWD}/src:/out" openapitools/openapi-generator-cli generate \
-    -i https://rest.bimdive.com \
+    -i $API_URI \
     -g typescript-fetch \
     -o /out \
     -p typescriptThreePlus=true,useSingleRequestParameter=true,supportsES6=true
